fix players not appearing in list (room foorm), reaplace vBox and scrollPane with the listView for players inside the room, add rest api for game/pool summary on gameServer, implement join room request on client and server

+++ Source/server/rabbit-game/src/Services/RabbitConnection.cs
@@ -7,6 +7,8 @@ namespace RabbitGameServer.Service
 	public class RabbitConnection : IRabbitConnection, IDisposable
 	{
 
+		// private const int DEFAULT_RABBIT_PORT = 5672;
+
 		private RabbitConfig config;
 
 		private IConnection connection;
@@ -21,7 +23,17 @@ namespace RabbitGameServer.Service
 
 			var factory = new ConnectionFactory();
 			factory.HostName = this.config.HostName;
-			factory.Port = this.config.Port;
+
+			factory.Port = (int)this.config.Port;
+			// if (this.config.Port != null)
+			// {
+			// 	factory.Port = (int)this.config.Port;
+			// }
+			// else
+			// {
+			// 	factory.Port = DEFAULT_RABBIT_PORT;
+			// }
+
 			factory.VirtualHost = this.config.VHost;
 			factory.UserName = this.config.UserName;
 			factory.Password = this.config.Password;
